retail/requirements.txt
asn1crypto==0.24.0
awscli==1.16.45
bcrypt==3.1.4
boto3==1.9.35
botocore==1.12.35
cffi==1.11.5
colorama==0.3.9
cryptography==2.3.1
Django==1.11.16
django-background-tasks==1.2.0
django-compat==1.0.15
docutils==0.14
enum34==1.1.6
futures==3.2.0
idna==2.7
ipaddress==1.0.22
jmespath==0.9.3
MySQL-python==1.2.5
numpy==1.15.4
pandas==0.23.4
paramiko==2.4.2
pyasn1==0.4.4
pycparser==2.19
PyMySQL==0.9.2
PyNaCl==1.3.0
python-dateutil==2.7.5
pytz==2018.5
PyYAML==3.13
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
urllib3==1.24
